+++ packages/expo-media-library/ios/MediaLibraryModule.swift
@@ -183,7 +183,6 @@ public class MediaLibraryModule: Module, PhotoLibraryObserverHandler {
           }
           let assets = getAssetsBy(assetIds: assetIds)
 
-          let collectionAssets = PHAsset.fetchAssets(in: collection, options: nil)
           let albumChangeRequest = PHAssetCollectionChangeRequest(for: collection, assets: assets)
           albumChangeRequest?.removeAssets(assets)
         } completionHandler: { success, _ in
@@ -457,9 +456,9 @@ public class MediaLibraryModule: Module, PhotoLibraryObserverHandler {
         self.allAssetsFetchResult = changeDetails.fetchResultAfterChanges
 
         if changeDetails.hasIncrementalChanges && !changeDetails.insertedObjects.isEmpty || !changeDetails.removedObjects.isEmpty {
-          var insertedAssets = [[String: Any]?]()
-          var deletedAssets = [[String: Any]?]()
-          var updatedAssets = [[String: Any]?]()
+          var insertedAssets = [[String: Any?]?]()
+          var deletedAssets = [[String: Any?]?]()
+          var updatedAssets = [[String: Any?]?]()
           let body: [String: Any] = [
             "hasIncrementalChanges": true,
             "insertedAssets": insertedAssets,